# docs/openapi/openapi.yaml
components:
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact: {}
  description: Deterministic, highly resilient backend infrastructure for the Lemni Payment-as-a-Service (PaaS) engine.
  title: LEMNI PaaS Core Engine API
  version: 1.0.0
openapi: 3.0.0
paths:
...
  /admin/api-keys:
    get:
      description: >-
        Returns a list of all API keys owned by the authenticated merchant, including their status and creation date.
        The secret part is never returned.
      operationId: ApiKeysController_listKeys
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  properties:
                    createdAt:
                      description: Key creation timestamp
                      example: '2026-07-04T10:00:00Z'
                      format: date-time
                      type: string
                    environment:
                      description: Key environment
                      enum:
                        - test
                        - live
                      example: test
                      type: string
                    id:
                      description: Key identifier
                      example: abc123def456
                      type: string
                    isActive:
                      description: Whether the key is active (false if revoked)
                      example: true
                      type: boolean
                  type: object
                type: array
          description: API keys retrieved successfully
        '401':
          description: Missing or invalid JWT token
      security:
        - bearer: []
      summary: List all API keys for the merchant
      tags:
        - merchant-dashboard/api-keys
    post:
      description: >-
        Creates and returns a new API key for merchant API authentication. The full key is returned only once - store it
        securely. Use test keys for development and live keys for production.
      operationId: ApiKeysController_generateKey
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                environment:
                  description: API key environment. Use "test" for development, "live" for production.
                  enum:
                    - test
                    - live
                  example: test
                  type: string
              required:
                - environment
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                properties:
                  keyId:
                    description: Key identifier for reference in dashboards
                    example: abc123def456
                    type: string
                  message:
                    description: Security reminder
                    example: Store this key safely. You will not be able to see it again.
                    type: string
                  rawKey:
                    description: Complete API key. Store this securely - you will not see it again.
                    example: sk_test_abc123def456_xyz789abc123def456xyz789abc123
                    type: string
                type: object
          description: API key successfully generated
        '400':
          description: Invalid environment value
        '401':
          description: Missing or invalid JWT token
      security:
        - bearer: []
      summary: Generate a new API key
      tags:
        - merchant-dashboard/api-keys
  /admin/api-keys/{id}:
    delete:
      description: >-
        Marks the specified API key as inactive immediately. All future requests using this key will be rejected with a
        403 Forbidden response.
      operationId: ApiKeysController_revokeKey
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    example: API key successfully revoked
                    type: string
                  success:
                    example: true
                    type: boolean
                type: object
          description: API key successfully revoked
        '401':
          description: Missing or invalid JWT token
        '404':
          description: API key not found or does not belong to this merchant
      security:
        - bearer: []
      summary: Revoke an API key
      tags:
        - merchant-dashboard/api-keys
...
  /auth/forgot-password:
    post:
      description: Generates and sends a 6-digit OTP code to the merchant email.
      operationId: AuthController_forgotPassword
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  description: Merchant email address
                  example: acme@example.com
                  format: email
                  type: string
              required:
                - email
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    description: Confirmation message
                    example: Reset code sent to acme@example.com (or email not found)
                    type: string
                type: object
          description: Reset code sent or email not found (for security, same message for both cases)
        '400':
          description: Invalid email format
      summary: Request password reset verification code
      tags:
        - merchant-dashboard/auth
  /auth/login:
    post:
      description: >-
        Authenticates merchant credentials and returns access token (1h expiry) and refresh token (7d expiry). Use
        access token for protected endpoints.
      operationId: AuthController_login
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  description: Registered merchant email
                  example: acme@example.com
                  format: email
                  type: string
                password:
                  description: Merchant account password
                  example: SecurePassword123
                  format: password
                  type: string
              required:
                - email
                - password
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    description: JWT access token (valid for 1 hour)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    type: string
                  merchant:
                    properties:
                      email:
                        format: email
                        type: string
                      id:
                        format: uuid
                        type: string
                      name:
                        type: string
                    type: object
                  refreshToken:
                    description: JWT refresh token (valid for 7 days)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    type: string
                type: object
          description: Login successful
        '401':
          description: Invalid email or password
      summary: Authenticate merchant and receive JWT tokens
      tags:
        - merchant-dashboard/auth
  /auth/logout:
    post:
      description: >-
        Notifies the server of logout. Client should immediately discard stored tokens. Tokens are stateless; this is a
        notification endpoint.
      operationId: AuthController_logout
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    example: Successfully logged out. Please discard tokens.
                    type: string
                type: object
          description: Successfully logged out
        '401':
          description: Missing or invalid access token
      security:
        - bearer: []
      summary: Logout (invalidate session)
      tags:
        - merchant-dashboard/auth
  /auth/refresh:
    post:
      description: >-
        Exchange a valid refresh token for a new access token (1h expiry). Use this before your access token expires to
        maintain an active session.
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                refreshToken:
                  description: Valid refresh token from login response
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  type: string
              required:
                - refreshToken
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  accessToken:
                    description: New JWT access token (valid for 1 hour)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                    type: string
                type: object
          description: Token refreshed successfully
        '401':
          description: Invalid or expired refresh token
      summary: Refresh access token using refresh token
      tags:
        - merchant-dashboard/auth
  /auth/reset-password:
    post:
      description: Resets the password of the merchant account verifying their current old password.
      operationId: AuthController_resetPassword
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                email:
                  description: The email address associated with the merchant account
                  example: acme@example.com
                  format: email
                  type: string
                newPassword:
                  description: The new password (minimum 8 characters)
                  example: NewSecurePassword123
                  format: password
                  minLength: 8
                  type: string
                oldPassword:
                  description: The current password
                  example: SecurePassword123
                  format: password
                  type: string
              required:
                - email
                - oldPassword
                - newPassword
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    description: Confirmation message
                    example: Password successfully reset
                    type: string
                  success:
                    example: true
                    type: boolean
                type: object
          description: Password successfully reset
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Invalid email, incorrect old password, or weak new password (minimum 8 characters)
                    type: string
                type: object
          description: Invalid email, incorrect old password, or weak new password
      summary: Reset password using old password verification
      tags:
        - merchant-dashboard/auth
        - merchant-dashboard/auth
  /auth/reset-password-with-token:
    post:
      description: Resets the password utilizing the short-lived reset token obtained from OTP verification.
      operationId: AuthController_resetPasswordWithToken
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              properties:
                newPassword:
                  description: New password (minimum 8 characters)
                  example: NewSecurePassword123
                  format: password
                  minLength: 8
                  type: string
                token:
                  description: Reset token from verify-reset-otp response (valid for 15 minutes)
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  type: string
              required:
                - token
                - newPassword
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  message:
                    description: Confirmation message
                    example: Password reset successfully
                    type: string
                  success:
                    description: Success flag
                    example: true
                    type: boolean
                type: object
          description: Password reset successfully
        '400':
          content:
            application/json:
              schema:
                properties:
                  error:
                    description: Error message
                    example: Invalid or expired token, or password does not meet requirements (minimum 8 characters)
                    type: string
                type: object
          description: Invalid/expired token or password criteria not met
      summary: Reset password using token
      tags:
        - merchant-dashboard/auth
...
  /auth/telegram-link:
    get:
      description: >-
        Returns a clickable Telegram link that opens the bot with merchant username pre-filled. Merchant just needs to
        click and the Telegram app opens with /start merchant_username ready to send.
      operationId: AuthController_getTelegramLink
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                properties:
                  telegramUrl:
                    description: Click this link to open Telegram with merchant username auto-filled
                    example: https://t.me/lemni_bot?start=acme-corp
                    format: uri
                    type: string
                type: object
          description: Telegram link generated successfully
        '400':
          description: Telegram bot not configured or merchant not found
        '401':
          description: Missing or invalid access token
      security:
        - bearer: []
      summary: Generate Telegram deep link for merchant support
      tags:
        - merchant-dashboard/auth
...
servers: []
...